+++ lib/src/digita_router_impl.dart
@@ -13,35 +13,107 @@ class Digita {
   /// Must be assigned to `navigatorKey` in your [MaterialApp].
   final GlobalKey<NavigatorState> navigatorKey = GlobalKey<NavigatorState>();
 
+  /// Ensure digita router is ready
+  ///
+  /// Assign (navigatorKey: digita.navigatorKey) to your MaterialApp
+  void ensureRouterIsReady() {
+    if (navigatorKey.currentState == null) {
+      debugPrint(
+          '[Digita] Navigator is not ready. Make sure you assigned [navigatorKey: digita.navigatorKey] to MaterialApp');
+      throw Exception('Navigator not ready');
+    }
+  }
+
   /// Push a new [Widget] page onto the navigation stack.
-  Future<dynamic>? goTo(Widget page) {
+  Future<Object?>? openPage(Widget page) {
+    ensureRouterIsReady();
     return navigatorKey.currentState?.push(
       MaterialPageRoute(builder: (_) => page),
     );
   }
 
-  /// Push a named route onto the navigation stack.
+  /// Pop the current page off the stack.
+  void closePage() {
+    navigatorKey.currentState?.pop();
+  }
+
+  /// Pop all pages off the stack until the root ('/') route.
+  void closeAllPages() {
+    navigatorKey.currentState?.popUntil((route) => route.isFirst);
+  }
+
+  /// Push a new [Widget] page onto the navigation stack.
   ///
+  /// Alias for [openPage].
+  Future<Object?>? goTo(Widget page) {
+    return openPage(page);
+  }
+
+  /// Push a named route onto the navigation stack.
   /// Optionally pass [arguments].
-  Future<dynamic>? goToNamed(String routeName, {Object? arguments}) {
+  Future<Object?>? openRoute(String routeName, {Object? arguments}) {
+    ensureRouterIsReady();
     return navigatorKey.currentState?.pushNamed(
       routeName,
       arguments: arguments,
     );
   }
 
-  /// Pop the current page off the stack.
+  /// Close the current route and go back to the previous
+  ///
+  /// Alias for [closePage].
+  void closeRoute() {
+    closePage();
+  }
+
+  /// Push a named route onto the navigation stack.
+  /// Optionally pass [arguments].
+  /// Alias for [openRoute].
+  Future<Object?>? goToRoute(String routeName, {Object? arguments}) {
+    ensureRouterIsReady();
+    return navigatorKey.currentState?.pushNamed(
+      routeName,
+      arguments: arguments,
+    );
+  }
+
+  /// Go back to the previous page.
+  ///
+  /// Alias for [closePage].
   void goBack() {
-    return navigatorKey.currentState?.pop();
+    closePage();
   }
 
   /// Pop pages until the specified [routeName] is reached.
   void goBackUntil(String routeName) {
-    return navigatorKey.currentState?.popUntil(ModalRoute.withName(routeName));
+    navigatorKey.currentState?.popUntil(ModalRoute.withName(routeName));
   }
 
   /// Replace the current page with a new [Widget] page.
-  Future<dynamic>? replaceWith(Widget page) {
+  Future<Object?>? replacePageWith(Widget page) {
+    return navigatorKey.currentState?.pushReplacement(
+      MaterialPageRoute(builder: (_) => page),
+    );
+  }
+
+  /// Replace the current page with a named route.
+  /// Optionally pass [arguments].
+  Future<Object?>? replaceRouteWith(String routeName, {Object? arguments}) {
+    return navigatorKey.currentState?.pushReplacementNamed(
+      routeName,
+      arguments: arguments,
+    );
+  }
+
+  /// ------------------
+  /// Deprecated methods
+  /// ------------------
+  ///
+  /// Replace the current page with a new [Widget] page.
+  ///
+  /// Deprecated: Use [replacePageWith] instead.
+  @Deprecated('Use replacePageWith instead')
+  Future<Object?>? replaceWith(Widget page) {
     return navigatorKey.currentState?.pushReplacement(
       MaterialPageRoute(builder: (_) => page),
     );
@@ -49,9 +121,21 @@ class Digita {
 
   /// Replace the current page with a named route.
   ///
-  /// Optionally pass [arguments].
-  Future<dynamic>? replaceNamed(String routeName, {Object? arguments}) {
+  /// Deprecated: Use [replaceRouteWith] instead.
+  @Deprecated('Use replaceRouteWith instead')
+  Future<Object?>? replaceNamed(String routeName, {Object? arguments}) {
     return navigatorKey.currentState?.pushReplacementNamed(
+      routeName,
+      arguments: arguments,
+    );
+  }
+
+  /// Push a named route onto the navigation stack.
+  ///
+  /// Deprecated: Use [goToRoute] instead.
+  @Deprecated('Use goToRoute instead')
+  Future<Object?>? goToNamed(String routeName, {Object? arguments}) {
+    return navigatorKey.currentState?.pushNamed(
       routeName,
       arguments: arguments,
     );